requirements.txt
fastapi
opencv-python
pillow
pytesseract
python-multipart
uvicorn
